LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.7_Punctuators/S7.7_A2_T8-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unrecognized token '\u002D'
S7.7_A2_T8

PASS Expected parsing failure

TEST COMPLETE
